requirements.txt
fastapi
uvicorn
google-generativeai
pandas
openpyxl
python-multipart
tabulate
matplotlib
seaborn
streamlit
requests